Update prepareUpdate helper to work with new OOP structure
